Zenject implementation in first level
Currently, the other levels are not playable, but the first level
is totally works.

Added Signals. Now in many classes C# events is replaced
by Signals, which allows to remove hard links between classes.

Settings are now subscribed to OnQuitSignal (which fires in
OnApplicationQuit), so the settings are now saved.

Added AudioManager, that now controls MusicPlayer and all SoundsPlayers

Added LevelInstaller and BaseInstaller, that replace LevelLoadedHandler
and BaseSceneLoadedHandler.

HintSystem was moved to Base scene. Hint steps are now stored
in LevelHintSteps, which is located in the Level scene.

Updated ProjectInstaller
Updated MenuInstaller